Remove add_grid_arguments_from_xarray
It's been completely replaced by parse_grid_arguments, and is only used
by a single test. Move that to be a test of parse_grid_arguments.
